feat: add error log to sentry
