+++ p2p_wallet/Services/TransactionHandler/TransactionHandler+Extensions.swift
@@ -262,24 +262,3 @@ extension TransactionHandler {
         }
     }
 }
-
-// MARK: - Helpers
-
-extension SolanaSwift.APIClientError: CustomNSError {
-    public var errorUserInfo: [String : Any] {
-        func getDebugDescription() -> String {
-            switch self {
-            case .cantEncodeParams:
-                return "Can not decode params"
-            case .invalidAPIURL:
-                return "Invalid APIURL"
-            case .invalidResponse:
-                return "Invalid response"
-            case .responseError(let response):
-                return response.message ?? "\(response)"
-            }
-        }
-        
-        return [NSDebugDescriptionErrorKey: getDebugDescription()]
-    }
-}
